third_party/blink/web_tests/external/wpt/trusted-types/TrustedTypePolicyFactory-createPolicy-cspTests-none.html

<!DOCTYPE html>
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<script src="./support/csp-violations.js"></script>
<meta http-equiv="Content-Security-Policy" content="trusted-types 'none'">
<meta http-equiv="Content-Security-Policy" content="object-src 'none'">
<body>
<script>
  promise_test(async t => {
    let violation = await trusted_type_violation_for(TypeError, _ =>
      window.trustedTypes.createPolicy('SomeName', { createHTML: s => s } )
    );
    assert_true(violation.originalPolicy.includes("trusted-types 'none'"));
  }, "Cannot create policy with name 'SomeName' - policy creation throws");

  promise_test(async t => {
    let violation = await trusted_type_violation_for(TypeError, _ =>
      window.trustedTypes.createPolicy('default', { createHTML: s => s } )
    );
    assert_true(violation.originalPolicy.includes("trusted-types 'none'"));
  }, "Cannot create policy with name 'default' - policy creation throws");
</script>
