LayoutTests/imported/w3c/web-platform-tests/trusted-types/TrustedTypePolicyFactory-createPolicy-cspTests-none-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'trusted-types'.


PASS Cannot create policy with name 'SomeName' - policy creation throws
PASS Cannot create policy with name 'default' - policy creation throws
